# Refactor get_degrees() to make it faster