--- a/src/ordering.jl
+++ b/src/ordering.jl
@@ -30,7 +30,8 @@ For a sum expression, returns the `get_degrees()` result for term with the highe
 See also `monomial_lt` and `lexlt`.
 """
 function get_degrees(expr)
-    res = _get_degrees(expr)
+    degs_cache = Dict()
+    res = _get_degrees(expr, degs_cache)
     if res isa AbstractVector
         return Tuple(res)
     else
@@ -38,44 +39,45 @@ function get_degrees(expr)
     end
 end
 
-function _get_degrees(expr)
+function _get_degrees(expr, degs_cache::AbstractDict)
     if issym(expr)
-        return ((Symbol(expr),) => 1,)
+        return get!(() -> ((Symbol(expr),) => 1,), degs_cache, expr)
     elseif iscall(expr)
         # operation-specific degree handling
-        return _get_degrees(operation(expr), expr)
+        return _get_degrees(operation(expr), expr, degs_cache)
     else
         return () # skip numbers and unsupported expressions
     end
 end
 
 # fallback for unsupported operation
-_get_degrees(::Any, expr) =
+_get_degrees(::Any, expr, degs_cache) =
     ((Symbol("zzzzzzz", hash(expr)),) => 1,)
 
 _getindex_symbol(arr, i) = Symbol(arr[i])
 
-function _get_degrees(::typeof(getindex), expr)
+function _get_degrees(::typeof(getindex), expr, degs_cache)
     args = arguments(expr)
-    @inbounds return (ntuple(Base.Fix1(_getindex_symbol, args), length(args)) => 1,)
+    @inbounds return get!(() -> (ntuple(Base.Fix1(_getindex_symbol, args), length(args)) => 1,),
+                degs_cache, expr)
 end
 
-function _get_degrees(::typeof(*), expr)
+function _get_degrees(::typeof(*), expr, degs_cache)
     args = arguments(expr)
     ds = sizehint!(Vector{Any}(), length(args))
     for arg in args
-        degs = _get_degrees(arg)
+        degs = _get_degrees(arg, degs_cache)
         append!(ds, degs)
     end
     return sort!(ds)
 end
 
-function _get_degrees(::typeof(+), expr)
+function _get_degrees(::typeof(+), expr, degs_cache)
     # among the terms find the best in terms of monomial_lt
     sel_degs = ()
     sel_degsum = 0
     for arg in arguments(expr)
-        degs = _get_degrees(arg)
+        degs = _get_degrees(arg, degs_cache)
         degsum = sum(last, degs, init=0)
         if (sel_degs == ()) || (degsum > sel_degsum) ||
            (degsum == sel_degsum && lexlt(degs, sel_degs))
@@ -85,10 +87,10 @@ function _get_degrees(::typeof(+), expr)
     return sel_degs
 end
 
-function _get_degrees(::typeof(^), expr)
+function _get_degrees(::typeof(^), expr, degs_cache)
     base_expr, pow_expr = arguments(expr)
     if pow_expr isa Number
-        @inbounds degs = map(_get_degrees(base_expr)) do (base, pow)
+        @inbounds degs = map(_get_degrees(base_expr, degs_cache)) do (base, pow)
             (base => pow * pow_expr)
         end
         if pow_expr < 0 && length(degs) > 1
@@ -99,23 +101,23 @@ function _get_degrees(::typeof(^), expr)
         return degs
     else
         # expression in the power argument is not supported
-        return _get_degrees(nothing, expr)
+        return _get_degrees(nothing, expr, degs_cache)
     end
 end
 
-function _get_degrees(::typeof(/), expr)
+function _get_degrees(::typeof(/), expr, degs_cache)
     nom_expr, denom_expr = arguments(expr)
     if denom_expr isa Number # constant denominator
-        return _get_degrees(nom_expr)
+        return _get_degrees(nom_expr, degs_cache)
     elseif nom_expr isa Number # constant nominator
-        @inbounds degs = map(_get_degrees(denom_expr)) do (base, pow)
+        @inbounds degs = map(_get_degrees(denom_expr, degs_cache)) do (base, pow)
             (base => -pow)
         end
         isa(degs, AbstractVector) || (degs = collect(degs))
         return sort!(degs)
     else
         # TODO expressions in both nom and denom are not yet supported
-        return _get_degrees(nothing, expr)
+        return _get_degrees(nothing, expr, degs_cache)
     end
 end