@AayushSabharwal 2025-03-28
Thanks for implementing this, the performance improvement is great. I have a couple concerns:
- It seems to have different behavior. The Symbolics reference test is failing
- Could `_get_degrees` be written so it doesn't dispatch on `op`? Since `BasicSymbolic` is type-unstable (out of necessity) it's always going to face runtime dispatch this way, and would be better off as an ifelse chain.