@alyst 2025-04-02
I've changed the sorting of negative degree:
 * for `lexlt()`: terms are sorted by descending *abs(deg)*, if equal, positive degree is first
 * for `monomial_lt()`: terms are sorted by ascending *abs(deg)*, if equal, positive degree is first

That should fix the *Symbolics.jl* integration tests (I've also added the failing expression to *SymbolicUtils.jl* unit tests).
The failures in *ModelingToolkit.jl* don't seem to be directly related to this PR. Probably some spurious numerical method failures.