@AayushSabharwal 2025-04-21
This looks good. I _think_ the Catalyst failure is possibly `structural_simplify` being temperamental. Now that benchmarks work, could you add your script above to the benchmark CI?