@alyst 2025-05-25
@AayushSabharwal Sorry for the long delay. Maybe it has something to do with `large_poly` var being new variable binding for the let block. I have changed that in the last commit.
But I suspect the real reason is that the PR is checked against the benchmarks defined in the default branch, not the PR itself, see how [benchpkg is run](https://github.com/JuliaSymbolics/SymbolicUtils.jl/blob/master/.github/workflows/benchmark_pr.yml#L41).
So for the `"get_degrees"` benchmark to show up in the PR reports, you need to accept this PR first.

I have checked that locally "get_degrees" will run as a part of `SUITE`:
```
julia> results = run(SUITE)
2-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "polyform" => 4-element BenchmarkTools.BenchmarkGroup:
          tags: []
          "easy_iszero" => Trial(35.500 μs)
          "simplify_fractions" => Trial(1.696 ms)
          "iszero" => Trial(1.240 ms)
          "isone" => Trial(0.001 ns)
  "overhead" => 7-element BenchmarkTools.BenchmarkGroup:
          tags: []
          "ruleset" => 3-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "noop:Int" => Trial(100.000 ns)
                  "noop:Sym" => Trial(100.000 ns)
                  "noop:Term" => Trial(4.000 μs)
          "simplify" => 7-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "randterm (/, *):serial" => Trial(211.300 μs)
                  "noop:Int" => Trial(200.000 ns)
                  "noop:Sym" => Trial(200.000 ns)
                  "randterm (/, *):thread" => Trial(248.300 μs)
                  "noop:Term" => Trial(45.400 μs)
                  "randterm (+, *):serial" => Trial(127.347 ms)
                  "randterm (+, *):thread" => Trial(149.125 ms)
          "acrule" => 5-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "noop:Int" => Trial(0.001 ns)
                  "noop:Sym" => Trial(0.001 ns)
                  "a+2+b" => Trial(800.000 ns)
                  "a+b" => Trial(200.000 ns)
                  "a+2" => Trial(800.000 ns)
          "large_poly" => 0-element BenchmarkTools.BenchmarkGroup:
                  tags: []
          "rule" => 3-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "noop:Int" => Trial(0.001 ns)
                  "noop:Sym" => Trial(0.001 ns)
                  "noop:Term" => Trial(0.001 ns)
          "substitute" => 3-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "a,b,c" => Trial(15.900 μs)
                  "a,b" => Trial(73.700 μs)
                  "a" => Trial(85.700 μs)
          "get_degrees" => 1-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "large_poly" => Trial(141.879 ms)

julia> results["overhead"]["get_degrees"]["large_poly"]
BenchmarkTools.Trial: 32 samples with 1 evaluation per sample.
 Range (min … max):  141.879 ms … 221.265 ms  ┊ GC (min … max): 0.00% … 17.56%
 Time  (median):     157.718 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   159.829 ms ±  16.505 ms  ┊ GC (mean ± σ):  2.88% ±  5.48%

    ██       ▄  ▁ ▁    ▁
  ▆▆██▆▆▁▁▆▁▁█▆▁█▆█▆▆▆▆█▁▁▁▁▆▁▁▁▁▁▁▆▁▆▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▆ ▁
  142 ms           Histogram: frequency by time          221 ms <

 Memory estimate: 43.92 MiB, allocs estimate: 1448503.
```